requirements.txt
Keras==2.3.1
aiohttp==3.7.3
pyzmq==19.0.2
numpy==1.19.4
pandas==1.1.3
opencv-python==4.1.2.30
matplotlib==3.3.2
gdown==3.12.2

editdistance==0.5.3
requests==2.25.0
cffi==1.14.4
PyQt5==5.10.1
future==0.18.2
Pillow==8.0.1
olefile==0.46
pip==20.3.3
wheel==0.35.1
pexpect==4.2.1
tornado==6.0.4
attrs==20.3.0
six==1.15.0
h5py==2.10.0
ipython==7.16.1
pytz==2020.1
tqdm==4.54.1
ipywidgets==7.5.1
notebook==6.1.4
PyYAML==5.3.1
idna==2.10
multidict==5.0.2
filelock==3.0.12
tensorflow==1.14.0
scipy==1.5.4
Cython==0.29.21
setuptools==50.3.2
python-dateutil==2.8.1
yarl==1.6.3
certifi==2020.11.8
chardet==3.0.4
urllib3==1.26.2
cryptography==3.1.1
simplejson==3.13.2
cycler==0.10.0
pygobject==3.26.1
ipykernel==5.3.4
Jinja2==2.11.2
pyparsing==2.4.7
kiwisolver==1.3.0
scikit-learn==0.23.2